backend/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
google-generativeai>=0.3.0
python-docx>=1.1.0
djangorestframework-simplejwt>=5.3.1
